# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/admin/donation-programs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all donation programs with cursor-based pagination and optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation Programs"
                ],
                "summary": "Get All Donation Programs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor (next page)",
                        "name": "next_cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor (prev page)",
                        "name": "prev_cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new donation program entry (requires authentication and proper role)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation Programs"
                ],
                "summary": "Create Donation Program",
                "parameters": [
                    {
                        "enum": [
                            "education",
                            "health",
                            "environment",
                            "social",
                            "disaster",
                            "humanity",
                            "other"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "CategoryEducation",
                            "CategoryHealth",
                            "CategoryEnvironment",
                            "CategorySocial",
                            "CategoryDisaster",
                            "CategoryHumanity",
                            "CategoryOther"
                        ],
                        "name": "category",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "end_date",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "name": "fund_target",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "start_date",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "active",
                            "draft",
                            "complete",
                            "expired"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "StatusActive",
                            "StatusDraft",
                            "StatusCompleted",
                            "StatusExpired"
                        ],
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Donation Program Cover Image",
                        "name": "cover_image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
...
        "/api/admin/donation-programs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information of a specific donation program",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation Programs"
                ],
                "summary": "Get Donation Program by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donation Program ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing donation program (requires authentication and proper role)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation Programs"
                ],
                "summary": "Update Donation Program",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donation Program ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "education",
                            "health",
                            "environment",
                            "social",
                            "disaster",
                            "humanity",
                            "other"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "CategoryEducation",
                            "CategoryHealth",
                            "CategoryEnvironment",
                            "CategorySocial",
                            "CategoryDisaster",
                            "CategoryHumanity",
                            "CategoryOther"
                        ],
                        "name": "category",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "end_date",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "name": "fund_target",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "start_date",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "active",
                            "draft",
                            "complete",
                            "expired"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "StatusActive",
                            "StatusDraft",
                            "StatusCompleted",
                            "StatusExpired"
                        ],
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Donation Program Cover Image",
                        "name": "cover_image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a donation program (requires authentication and proper role)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation Programs"
                ],
                "summary": "Delete Donation Program",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
...
        "/api/donation-programs": {
            "get": {
                "description": "Retrieve a list of published (active) donation programs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation Programs"
                ],
                "summary": "List Published Donation Programs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor (next page)",
                        "name": "next_cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor (prev page)",
                        "name": "prev_cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
...
        "/api/donation-programs/{slug}": {
            "get": {
                "description": "Get detailed information of a specific published donation program",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation Programs"
                ],
                "summary": "Get Published Donation Program by Slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donation Program Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pkg.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "validation": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}